Scrape packit.dide.ic.ac.uk
Rather than hardcoding a list of targets to scrape, the configuration
uses an HTTP-based service discovery. The packit.dide.ic.ac.uk machine
is configured to expose a JSON document on port 9000. This will allow
Prometheus to automatically discover new endpoints as we add new
instances.

Currently, the services being exposed are one node_exporter for the
whole machine, and one outpack_server for each instance hosted on the
server. Packit does not yet have a metrics endpoint.