Use the stock nginx image instead of monitor-proxy.
Having a separate repository makes any changes to the nginx
configuration an annoying multi-step process. The two repositories, this
one and monitor-proxy, are so tightly coupled there is little benefit to
keeping them separate. The monitor-proxy repository is nothing more than
an `nginx.conf` file and some build scaffolding.

The `nginx.conf` file was copied almost verbatim from the monitor-proxy,
with just a few tweaks to avoid unnecessary templating. The
`dhparam.pem` was copied as-is.