Fix the relabel config for Packit.
The relabel rule was trying to rewrite the instance label, but was not
working. It seems as though at the point where it runs there is no
instance label yet. It only gets set later in the process to
`__address__` if not set anywhere else.

Using `__address__` directly fixes that and makes the relabel rule work
as intended.