Add scraping of wpia-hn's redis instance.
This will help keep an eye of the size and growth of the database, as
well as queries latency. I was initially hoping to monitor error rates
as well, but it seems our Redis instance is fairly old and doesn't
expose these statistics.

This uses redis-exporter as a bridge between redis and prometheus.
In theory a single instance of the exporter can be used to monitor any
number of redis server's, although at the moment only the wpia-hn is
configured. We may want to add hint and/or daedalus' servers in the
future.

It seems redis has native support for prometheus in [its enterprise
offering][redis-metrics], but not something we get to use.

[redis-metrics]: https://redis.io/docs/latest/integrate/prometheus-with-redis-enterprise/prometheus-metrics-definitions/