Monitor and alert about Packit configuration drift.
While the packit servers are configured declaratively, it is possible
for someone to deploy from a branch / their local copy instead of from
main.

We want to ensure servers follow the upstream configuration as closely
as possible. This introduces a metric exposing the SHA of the upstream
main branch. This global metric can be cross-referenced against the
existing `nixos_configuration_info` exposed by each machine to trigger
an alert.

The metric for the GitHub repository is implemented as a short Go
script. The script exposes and HTTP endpoint with two query parameters,
`owner` and `repo`, and returns metrics for all branches of the
repository. The script is simple enough that it is embedded directly in
here instead of living in its own repository.

To give us some flexiblity in development and avoid superfluous alerts,
the alert only triggers after a long period of time, currently one week.
This allows temporary configuration drift, as long as we don't forget
about it and eventually reconcile it.