Allow anonymous access to Grafana.
We have a shared password (stored in Vault) that is used to restrict
access to Grafana. Given that Prometheus, the underlying datasource,
already allows unauthenticated access to the metrics, putting access
restrictions on Grafana doesn't protect anything and makes sharing
dashboards with other people tedious.

Modifying dashboards will still require logging in. A "Sign in" button
is present in the top-right of the webpage.

https://grafana.com/docs/grafana/v11.4/setup-grafana/configure-security/configure-authentication/#anonymous-authentication