Reduce the frequency of PackitConfigDrift alerts.
These are not time sensitive or of a high severity. We only have those
alerts to avoid forgetting to deploy. Once a week notice is enough (it
already takes a week to fire the first time).

For this to work we need to increase alertmanager's data retention
period, now at two weeks. The default value was 120h.