Improve and simplify the Slack alert template.
Alerts are now grouped by type. They include three buttons at the
bottom, linking to the query, in Prometheus, that triggered the alert,
to the page to create a new silence with the common fields already
populated and to the knowledge base article about monitoring.

I've removed the distinction between the hint slack channel and the rest
of the alerts. No one was ever looking at the hint channel.