Exclude cifs mountpoints from the disk space alerts.
CIFS mounts represent network drives. These are managed externally and
do not represent the health of the individual machines.

While it might be useful to have an eye on these drives, if we do it we
should do it in a consistent way, not just on the drives that we happen
to have mounted.

For reference the Climate drive (mounted on GPU-02) is currently at 83%
occupation, and the Q drive (mounted on GPU-01) is just under 80%.