Scrape the Windows HPC cluster.
We now run windows_exporter on all of the Windows HPC nodes, both
workers and headnodes.

The cluster nodes are not directly reachable by Prometheus, since they
live inside of a NAT. We've installed a simple nginx-based reverse-proxy
on the headnode, which exposes `/metrics/<ip>` as a way to scrape the
metrics of a particular cluster machine.

Done a bit of refactor to unify the buildkite and HPC inventories. I've
also modified some of the existing scrapes (eg. node_exporter and
DCGM) to remove the port number from their instance label.