Remove last remnants of AWS configuration.
We used to connect to AWS to fetch metrics from barman and bb8, but this
had been broken for a while and was removed from the scrape rules some
time ago. We still had some code left in place to fetch AWS credentials
from Vault, which is not necessary anymore.